+++ playground/src/main/kotlin/io/bkbn/lerasium/playground/domain/BookDomain.kt
@@ -3,6 +3,7 @@ package io.bkbn.lerasium.playground.domain
 import io.bkbn.lerasium.core.Domain
 import io.bkbn.lerasium.rdbms.Table
 import io.bkbn.lerasium.api.Api
+import io.bkbn.lerasium.rdbms.Unique
 
 @Domain("Book")
 private sealed interface BookDomain {
@@ -12,7 +13,10 @@ private sealed interface BookDomain {
 }
 
 @Table
-private interface BookTable : BookDomain
+private interface BookTable : BookDomain {
+  @Unique
+  override val isbn: String
+}
 
 @Api
 private interface BookApi : BookDomain